When aliasing a bitmap, use smallest available limit.
The original implementation of the bitmap aliasing routine chose the
conservative max value of the heap limit.  This is perfect for the
zygote but is oversized for application heaps that are not anywhere
near full.  Now the code consults the live bitmap and will use its max
value if it is smaller than the heap limit.

Change-Id: I7cf223efdeaed318922a8a5f9e147092e539da6c