Add class checks in AllocObject / NewObject.
The JNI spec says the functions throw InstantiationException if you
pass in an abstract or interface class, and that you "must not" try to
allocate an array class with this mechanism.  These are now checked.
We also forbid creation of a Class object with these calls, since
that should only be done by DefineClass.

Also, added appropriate "is class initialized" code for the A/V
variants of NewObject, which were strangely lacking them.

Change-Id: I8dde7a812265cfb16b1765695c652e98c9de0840