Fix build breakage caused by the countInstancesOfClass change.
There is an apparent visibility skew between passion-eng and sim-eng
builds.  On sim-eng, the contents of Heap.h are globally visible but
on passion-eng this is not the case.  To resolve this issue, I have
moved the offending declarations to alloc/Alloc.h which is directly
exported from Dalvik.h.

Change-Id: Iaf6f167964f81f24fa0817aa4a84b9fb4ff9a9a0