Fix a DDMS thread monitoring crash.
We changed the order of events during thread attaches, which means
it's now possible to have Thread structs with null threadObj fields.
This bit of code didn't change with the times, which made it possible
for a process being watched by DDMS to crash if we grabbed the thread
state at just the wrong time.

Change-Id: I3cd5501e99278d68897f8e50f316f8a37620ca0b