Remove an unnecessary finger check when re-marking roots.
The finger check, a less than check, is meaningful only during the
initial root marking.  After root marking, the finger is set to the
end of the virtual address space and the check always succeeds.  As
such, this check is superfluous during all other phases of garbage
collection as it always suceeds.

Change-Id: Ie4e5bc1fb57b41e484cb6d1081c9a874fa6effbc