Improve dexdump temp file handling.
When dexdump is handed a zip or APK, it has to unpack classes.dex to a
temp file.  It currently tries /tmp and /sdcard, neither of which are
likely to exist on a Windows system.

We now print a warning when those directories don't exist, and try to
create the temp file in the current directory.

Bug 3125957.

Change-Id: I7724ce48c35f22ed4b5116dfe5783243107df6a1