Clean up dx test 085.
The jasmin source had a bug which dx didn't previously catch, which is
now fixed, and I also removed a superfluous <init> method. With that
cleaned up, the test now fails but in a different way, which I believe
is indicative of a real problem, where it looks like a known constant
value is not getting properly propagated.

Change-Id: I3d31045eecc6d71cff2bca98c948dc58371d35a1