Rework the bytecode format file and parser.
This change adds a set of "format" commands to the bytecode.txt file
and prefixes each bytecode definition with an "op" keyword. The new
parser script understands both commands (though it mostly ignores the
"format" ones) and is now a little less wedded to the idea that there
can only be 256 opcodes.

This all is in preparation for (a) adding new opcodes, and (b) making
the opcode/instruction fitter a little more efficient.

Change-Id: I1b4e7274da06b197536823148d91f426d049ec19