Fix for Issue 3161543 - Remove stray line of code in JIT
During Change 22370 "Thumb/Thumb2 instruction selection rework"
line 1689 of Codegen.c was inadvertently duplicated and deposited
in the unrelated function genPuntToInterp() at line 1732.

genPuntToInterp() is rarely used and half of the time (if the
roll-back Dalvik PC was word-aligned) the error would have been
benign.  However, if the Dalvik PC is not word-aligned, the result
would be a seg fault.

The fix is to remove the line of code.

Change-Id: Ib2ffb80059a639923c342a36797eb6c4995c5218